permissao da entrada como visitante

+++ inc/cabecalho.inc.php
@@ -18,32 +18,36 @@ if (strpos($_SERVER['REQUEST_URI'], 'index.php') !== false) {
     if($horario > 17 && $horario <= 23){
         $saudacao = "Boa noite";
     }
-    if (strpos($_SERVER['REQUEST_URI'], 'sobre.php') == false) {
-        echo "<h3> $saudacao " . $_SESSION['nome'] . "!";
-        echo "<br>Reserva de ambientes - 2024</h3><br>";
-    }
-    echo "<a href=\"Logout.php\">Logout</a> <br>";
-    echo "<a href=\"ambiente.php\">Home</a> <br>";
-    echo "<a href=\"minhasreservas.php\">Minhas Reservas</a><br>";
-    if($_SESSION['admin'] == 'sim'){
-        echo "
-        <details>
-            <summary>Ambiente</summary>
-            <ul>
-            <li><a href=\"criarambientes.php\">Criar Ambiente</a></li>
-            <li><a href=\"excluirambiente.php\">Excluir Ambiente</a></li>
-            </ul>
-        </details>
-        ";
-        echo "
-        <details>
-            <summary>Usuário</summary>
-            <ul>
-            <li><a href=\"criarusuario.php\">Criar Usuário</a></li>
-            <li><a href=\"excluirusuario.php\">Excluir Usuário</a></li>
-            </ul>
-        </details>
-        ";
-        echo "<a href=\"todasasreservas.php\">Todas as reservas</a> <br>";
+    if($_SESSION['visitante']){
+        echo "<a href=\"Logout.php\">Logout</a> <br>";
+    } else {
+        if (strpos($_SERVER['REQUEST_URI'], 'sobre.php') == false) {
+            echo "<h3> $saudacao " . $_SESSION['nome'] . "!";
+            echo "<br>Reserva de ambientes - 2024</h3><br>";
+        }
+        echo "<a href=\"Logout.php\">Logout</a> <br>";
+        echo "<a href=\"ambiente.php\">Home</a> <br>";
+        echo "<a href=\"minhasreservas.php\">Minhas Reservas</a><br>";
+        if($_SESSION['admin'] == 'sim'){
+            echo "
+            <details>
+                <summary>Ambiente</summary>
+                <ul>
+                <li><a href=\"criarambientes.php\">Criar Ambiente</a></li>
+                <li><a href=\"excluirambiente.php\">Excluir Ambiente</a></li>
+                </ul>
+            </details>
+            ";
+            echo "
+            <details>
+                <summary>Usuário</summary>
+                <ul>
+                <li><a href=\"criarusuario.php\">Criar Usuário</a></li>
+                <li><a href=\"excluirusuario.php\">Excluir Usuário</a></li>
+                </ul>
+            </details>
+            ";
+            echo "<a href=\"todasasreservas.php\">Todas as reservas</a> <br>";
+        }
     }
 }
